pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "automic-monitor"
version = "0.2.0"
description = "Automic Workflow Monitoring and Processing System"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Health Partners Plans", email = "tech@healthpartners.com"}
]

dependencies = [
    # Core Dependencies
    "requests>=2.32.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.2",
    
    # Data Validation
    "pydantic>=2.12.5",
    "pydantic-settings>=2.0.0",
    "PyYAML>=6.0",
    
    # Database
    "pyodbc>=5.3.0",
    "sqlalchemy-access>=2.0.3",
    
    # File Processing
    "pyx12>=2.3.3",
    "edi-835-parser>=1.8.0",
    
    # Utilities
    "tabulate>=0.10.0",
    "openpyxl>=3.1.5",
    "setuptools<81",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]

[project.urls]
Homepage = "https://github.com/healthpartners/automic-monitor"
Documentation = "https://github.com/healthpartners/automic-monitor/wiki"
Repository = "https://github.com/healthpartners/automic-monitor"
Issues = "https://github.com/healthpartners/automic-monitor/issues"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/site-packages/*",
]
